Updating analysis functions notebook name, adjusting financial and other functions for NY run
