+++ batch_job_yamls/dgen-batch-job-mid-large-states.yaml
@@ -2,8 +2,8 @@ logsPolicy:
   destination: CLOUD_LOGGING
 
 taskGroups:
-  - taskCount: "12"               # one task per state
-    parallelism: "5"             # run all states in parallel
+  - taskCount: "1"               # one task per state
+    parallelism: "1"             # run all states in parallel
     schedulingPolicy: AS_SOON_AS_POSSIBLE
 
     taskSpec:
