+++ build_and_submit.sh
@@ -7,4 +7,4 @@ docker buildx build --platform linux/amd64 \
   --push .
 
 # Submit job
-./submit_all.sh
+./submit_one.sh
