+++ dgen_os/input_data/pv_plus_batt_prices/pv_plus_batt_prices_FY23_NY_policy.csv
@@ -1,0 +1,38 @@
+year,tech,system_capex_per_kw_res,system_capex_per_kw_nonres,batt_capex_per_kwh_res,batt_capex_per_kw_res,batt_capex_per_kwh_nonres,batt_capex_per_kw_nonres,linear_constant_res,linear_constant_nonres,batt_om_per_kw_res,batt_om_per_kwh_res,batt_om_per_kw_nonres,batt_om_per_kwh_nonres,batt_replace_frac_kw,batt_replace_frac_kwh
+2014,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2015,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2016,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2017,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2018,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2019,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2020,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2021,"pv,batt",2947.143747,1906.302812,717.9439627,554.8244171,234.6229377,327.5022644,6657.420662,397856.729,92.02921141,0,48.22721409,0,0,1
+2022,"pv,batt",2947.143747,1906.302812,597.3337942,646.4052909,264.3236332,957.8767111,7834.848561,215911.2606,92.66773721,0,59.37558362,0,0,1
+2023,"pv,batt",2841.725433,1847.959542,547.1023294,624.9827415,241.2616414,893.1255083,7575.193461,203028.735,87.69443143,0,54.91383247,0,0,1
+2024,"pv,batt",2736.307119,1789.616271,499.1649605,606.8969981,219.1634781,837.6016449,7355.982599,192207.1957,83.15014798,0,50.86502208,0,0,1
+2025,"pv,batt",2630.888805,1731.273,453.0136393,591.454816,208.7174523,803.0994457,7168.813405,184658.529,78.94378988,0,48.64333675,0,0,1
+2026,"pv,batt",4030,1672.929729,431.0272216,579.1630353,197.281737,779.3774908,7019.829103,180287.0818,76.51742275,0,46.72457271,0,0,1
+2027,"pv,batt",3980,1614.586459,411.6096429,567.9703857,186.8497318,755.8767993,6884.16698,175839.5747,74.34569722,0,44.9085421,0,0,1
+2028,"pv,batt",3920,1556.243188,395.1513059,557.7631993,177.2561349,733.8267751,6760.449305,171643.9905,72.44328313,0,43.2231158,0,0,1
+2029,"pv,batt",3800,1497.899917,379.6304541,547.3616117,168.587596,713.8006259,6634.375362,167824.6981,70.58282048,0,41.69647101,0,0,1
+2030,"pv,batt",3620,1439.556647,365.6123506,537.8433086,160.6218448,694.1035176,6519.007397,163995.3372,68.89189161,0,40.24791147,0,0,1
+2031,"pv,batt",3420,1381.213376,361.0421962,531.1202673,158.6140717,685.4272236,6437.519804,161945.3955,68.03074296,0,39.74481257,0,0,1
+2032,"pv,batt",3270,1322.870105,356.4720418,524.3972259,156.6062986,676.7509297,6356.032212,159895.4538,67.16959432,0,39.24171368,0,0,1
+2033,"pv,batt",3180,1264.526834,351.9018874,517.6741845,154.5985256,668.0746357,6274.544619,157845.5121,66.30844567,0,38.73861479,0,0,1
+2034,"pv,batt",3130,1206.183564,347.331733,510.9511432,152.5907525,659.3983417,6193.057027,155795.5703,65.44729703,0,38.23551589,0,0,1
+2035,"pv,batt",3090,1147.840293,342.7615786,504.2281018,150.5829795,650.7220478,6111.569434,153745.6286,64.58614838,0,37.732417,0,0,1
+2036,"pv,batt",3030,1128.690643,338.1914243,497.5050605,148.5752064,642.0457538,6030.081842,151695.6869,63.72499974,0,37.22931811,0,0,1
+2037,"pv,batt",2980,1109.540994,333.6212699,490.7820191,146.5674333,633.3694598,5948.594249,149645.7452,62.86385109,0,36.72621921,0,0,1
+2038,"pv,batt",2920,1090.391344,329.0511155,484.0589778,144.5596603,624.6931659,5867.106657,147595.8035,62.00270245,0,36.22312032,0,0,1
+2039,"pv,batt",2870,1071.241695,324.4809611,477.3359364,142.5518872,616.0168719,5785.619064,145545.8618,61.1415538,0,35.72002143,0,0,1
+2040,"pv,batt",2810,1052.092045,319.9108067,470.612895,140.5441142,607.3405779,5704.131472,143495.9201,60.28040516,0,35.21692253,0,0,1
+2041,"pv,batt",1000,1032.942396,315.3406524,463.8898537,138.5363411,598.6642839,5622.64388,141445.9783,59.41925651,0,34.71382364,0,0,1
+2042,"pv,batt",1000,1013.792746,310.770498,457.1668123,136.528568,589.98799,5541.156287,139396.0366,58.55810787,0,34.21072475,0,0,1
+2043,"pv,batt",1000,994.6430963,306.2003436,450.443771,134.520795,581.311696,5459.668695,137346.0949,57.69695922,0,33.70762585,0,0,1
+2044,"pv,batt",1000,975.4934467,301.6301892,443.7207296,132.5130219,572.635402,5378.181102,135296.1532,56.83581058,0,33.20452696,0,0,1
+2045,"pv,batt",1000,956.3437971,297.0600348,436.9976883,130.5052489,563.9591081,5296.69351,133246.2115,55.97466193,0,32.70142807,0,0,1
+2046,"pv,batt",1000,937.1941476,292.4898804,430.2746469,128.4974758,555.2828141,5215.205917,131196.2698,55.11351329,0,32.19832917,0,0,1
+2047,"pv,batt",1000,918.044498,287.9197261,423.5516055,126.4897027,546.6065201,5133.718325,129146.3281,54.25236464,0,31.69523028,0,0,1
+2048,"pv,batt",1000,898.8948484,283.3495717,416.8285642,124.4819297,537.9302261,5052.230732,127096.3863,53.391216,0,31.19213139,0,0,1
+2049,"pv,batt",1000,879.7451988,278.7794173,410.1055228,122.4741566,529.2539322,4970.74314,125046.4446,52.53006735,0,30.68903249,0,0,1
+2050,"pv,batt",1000,860.5955492,274.2092629,403.3824815,120.4663836,520.5776382,4889.255547,122996.5029,51.66891871,0,30.1859336,0,0,1
