+++ state_input_csvs/mid_large_states.csv
@@ -1,12 +1,1 @@
-IL,Illinois
-GA,Georgia
-PA,Pennsylvania
-OH,Ohio
 VA,Virginia
-NC,North Carolina
-MI,Michigan
-MO,Missouri
-TN,Tennessee
-KY,Kentucky
-IN,Indiana
-MN,Minnesota
