+++ state_input_csvs/mid_large_states_test.csv
@@ -1,12 +1,1 @@
-IN,Indiana
-KY,Kentucky
-IL,Illinois
-TN,Tennessee
-NC,North Carolina
-PA,Pennsylvania
-GA,Georgia
-MN,Minnesota
-VA,Virginia
-MO,Missouri
-OH,Ohio
-MI,Michigan
+NY,New York
