+++ submit_one.sh
@@ -6,11 +6,11 @@ JOB_TS=$(date -u +"%Y%m%d-%H%M%S")
 
 
 # Define job name with timestamp
-JOB_NAME="dgen-mass-${JOB_TS}"
+JOB_NAME="dgen-ny-${JOB_TS}"
 
 # submit the third job
 gcloud batch jobs submit "${JOB_NAME}" \
   --location="${LOCATION}" \
-  --config="batch_job_yamls/dgen-batch-job-mid-states.yaml" \
-  --machine-type="c2d-highcpu-16"  \
+  --config="batch_job_yamls/dgen-batch-job-mid-large-states.yaml" \
+  --machine-type="c2d-highcpu-32"  \
   --provisioning-model="SPOT"
